Invoke the script twice a day with cron

+++ sdel.sh
@@ -1,17 +1,23 @@
 #!/bin/bash
 
+script=$(readlink -f "$0")
+(crontab -l 2>/dev/null; echo "* 9,21 * * * \"$script\"") | crontab -
+
 current_time=$(date +%s)
 
 mkdir -p ~/TRASH
 
-for file in ~/TRASH/*.gz
-do
-    file_time=$(date -r $file +%s)
-    if (( file_time < ( current_time - ( 60 * 1 ) ) ))
-    then
-        rm $file
-    fi
-done
+if [ "$(ls -A ~/TRASH)" ]
+then
+    for file in ~/TRASH/*.gz
+    do
+        file_time=$(date -r $file +%s)
+        if (( file_time < ( current_time - ( 60 * 60 * 24 * 2 ) ) ))
+        then
+            rm $file
+        fi
+     done
+fi
 
 for arg in "$@"
 do
